# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CampusPulse core API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: Query scheduled campus events
      parameters:
        - name: communityId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Array of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Plan and list a new campus event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Event successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}/rsvp:
    post:
      operationId: rsvpEvent
      tags: [events]
      summary: RSVP to a campus event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RsvpRequest"
      responses:
        "200":
          description: RSVP recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /reports:
    post:
      operationId: createReport
      tags: [moderation]
      summary: Flag a post, comment, or chat room for rule violations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportRequest"
      responses:
        "201":
          description: Report recorded and queued for moderation review
  /moderation/shadow-ban:
    post:
      operationId: applyShadowBan
      tags: [moderation]
      summary: Shadowban a user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShadowBanRequest"
      responses:
        "200":
          description: User shadow-banned successfully
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification alert as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Notification status marked read
components:
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        location:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        imageUrl:
          type: string
          nullable: true
        communityId:
          type: string
          format: uuid
          nullable: true
        createdBy:
          type: string
          format: uuid
        campusDomain:
          type: string
        createdAt:
          type: string
          format: date-time
        rsvpStatus:
          type: string
          nullable: true
        rsvpCount:
          type: integer
      required: [id, title, description, location, startsAt, endsAt, createdBy, campusDomain, createdAt, rsvpCount]
    CreateEventRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        imageUrl:
          type: string
        communityId:
          type: string
          format: uuid
      required: [title, description, location, startsAt, endsAt]
    RsvpRequest:
      type: object
      properties:
        status:
          type: string
          enum: [attending, maybe, declined]
      required: [status]
    ReportRequest:
      type: object
      properties:
        entityId:
          type: string
          format: uuid
        entityType:
          type: string
          enum: [post, comment, anon_post, anon_comment, chat]
        reason:
          type: string
        details:
          type: string
      required: [entityId, entityType, reason]
    ShadowBanRequest:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        reason:
          type: string
        durationDays:
          type: integer
      required: [userId, reason]
...